+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 ### Added
 - Permissions.IsUserInGroup() to see if a user is in a group
 - TestEZ for testing
+- Functionality for using negated permission nodes
+- Internal function isNodeNegated() to check for negated permission nodes
 ### Changed
 - Permissions.Init() to take a table as a config rather then ModuleScript
+- Moonwave documentation typos fixed
 ### Removed
 - Outdated test config
 - Changelog unnecessary tags